Fix 64 bit compiler warnings in the stackless module.
git-svn-id: http://svn.python.org/projects/stackless/branches/release25-maint@55313 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Stackless/module/channelobject.c
@@ -921,7 +921,7 @@ channel_setstate(PyObject *self, PyObject *args)
 	PyObject *lis;
 	int flags, balance;
 	int dir;
-	int i, n;
+	Py_ssize_t i, n;
 
 	if (!PyArg_ParseTuple(args, "iiO!:channel",
 			      &balance,
